+++ data/changelog.md
@@ -1,3 +1,4 @@
+* **New:** You can now `import` data that you've previously `export`ed!
 * **Bug:** Accessing the command history using the up and down arrows will no
   longer wrap around when you reach the beginning or end of the history.
 * **Bug:** Clicking the right mouse button no longer runs commands or changes